requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
opencv-python
pillow
tqdm
streamlit
fastapi
uvicorn
pyyaml